-moved Icons styles + variables to teams theme
