+++ packages/react/src/themes/base/componentStyles.ts
@@ -1,8 +1,6 @@
 export { default as Flex } from './components/Flex/flexStyles'
 export { default as FlexItem } from './components/Flex/flexItemStyles'
 
-export { default as Icon } from './components/Icon/iconStyles'
-
 export { default as Toolbar } from './components/Toolbar/toolbarStyles'
 export { default as ToolbarRadioGroup } from './components/Toolbar/toolbarRadioGroupStyles'
