+++ packages/react/src/themes/base/componentVariables.ts
@@ -1,6 +1,4 @@
 export { default as Flex } from './components/Flex/flexVariables'
 export { default as FlexItem } from './components/Flex/flexItemVariables'
 
-export { default as Icon } from './components/Icon/iconVariables'
-
 export { default as Grid } from './components/Grid/gridVariables'
