--- packages/react/src/themes/base/components/Icon/iconNames.ts
@@ -1,22 +1,0 @@
-import { ThemeIcons, ThemeIconSpec } from '../../../types'
-
-const fontIcon = (content: string): ThemeIconSpec => ({
-  icon: { content: `'\\${content}'` },
-})
-
-export const icons: ThemeIcons = {
-  'stardust-checkmark': fontIcon('2713'),
-  'stardust-circle': fontIcon('25CF'),
-  'stardust-close': fontIcon('2715'),
-  'stardust-arrow-down': fontIcon('25BE'),
-  'stardust-arrow-end': fontIcon('25B8'),
-  'stardust-arrow-up': fontIcon('25B4'),
-  'stardust-menu-arrow-down': fontIcon('25BE'),
-  'stardust-menu-arrow-end': fontIcon('25B8'),
-  'stardust-pause': fontIcon('25B6'),
-  'stardust-play': fontIcon('23F8'),
-}
-
-export const emptyIcon: ThemeIconSpec = { icon: { content: '?' } }
-
-export default (name: string): ThemeIconSpec => icons[name] || emptyIcon
