+++ packages/react/src/themes/base/index.ts
@@ -2,11 +2,9 @@ import { ThemeInput } from '../types'
 import * as siteVariables from './siteVariables'
 import * as componentVariables from './componentVariables'
 import * as componentStyles from './componentStyles'
-import { icons } from './components/Icon/iconNames'
 
 export default {
   siteVariables,
-  icons,
   componentVariables,
   componentStyles,
 } as ThemeInput
