--- packages/react/src/themes/font-awesome/componentStyles.ts
@@ -1,1 +1,0 @@
-export { default as Icon } from './components/Icon/iconStyles'
