+++ packages/react/src/themes/font-awesome/index.ts
@@ -1,8 +1,6 @@
 import { ThemeInput } from '../types'
-import * as componentStyles from './componentStyles'
 import { default as icons } from './components/Icon/iconNames'
 
 export default {
   icons,
-  componentStyles,
 } as ThemeInput
